Feature/accounts/server (#28)
* Add Account context

Add the base for accounts context, which will handle user information,
roles and credentials.

* Add credentials on user create

Email and user password is required on user create.

* Add role migration and schema

A user can have many roles such as user or Admin. We can keep track
of these in a separate roles table. In a future commit on user or admin
create the correct role will be assigned.