register push notification token (#97)
* WIP

* move this

* feat: 🎸 Register device token

Allow users to register their device tokens in the data store. These
will later be used when sending push notifications to users devices when
for instance a plant is scheduled to be watered.

* fix: update array of devices if exists

When adding devices we can first try to update an existing document. If it does not exist we create
a new one.